requirements.txt
asgiref==3.3.4
attrs==20.3.0
Django==2.2.9
django-debug-toolbar==3.2.1
django-extensions==3.1.3
iniconfig==1.1.1
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==3.8.0
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
toml==0.10.2
